fix(6-unable-to-read-workflow-in-webp-image-previous-commit-works-fine): tidy fns
tidy fns
