+++ app/utils/utils.ts
@@ -1,0 +1,7 @@
+
+export function removeExt(f: string) {
+  if (!f) return f;
+  const p = f.lastIndexOf(".");
+  if (p === -1) return f;
+  return f.substring(0, p);
+}
